+++ tests/specs/unknown/unknown.spec.js
@@ -4,7 +4,7 @@ describe('Schema with $refs to unknown file types', function() {
   var windowOnError, testDone;
 
   beforeEach(function() {
-    // Some old Webkit browsers throw an error when downloading zero-byte files.
+    // Some browsers throw an error when downloading zero-byte files.
     windowOnError = global.onerror;
     global.onerror = function() {
       testDone();
@@ -19,14 +19,15 @@ describe('Schema with $refs to unknown file types', function() {
   it('should parse successfully', function(done) {
     testDone = done;
     var parser = new $RefParser();
-    parser
-      .parse(path.rel('specs/unknown/unknown.yaml'))
+
+    parser.parse(path.rel('specs/unknown/unknown.yaml'))
       .then(function(schema) {
         expect(schema).to.equal(parser.schema);
         expect(schema).to.deep.equal(helper.parsed.unknown.schema);
         expect(parser.$refs.paths()).to.deep.equal([path.abs('specs/unknown/unknown.yaml')]);
         done();
-      });
+      })
+      .catch(done);
   });
 
   it('should resolve successfully', function(done) {
@@ -43,8 +44,8 @@ describe('Schema with $refs to unknown file types', function() {
   it('should dereference successfully', function(done) {
     testDone = done;
     var parser = new $RefParser();
-    parser
-      .dereference(path.rel('specs/unknown/unknown.yaml'))
+
+    parser.dereference(path.rel('specs/unknown/unknown.yaml'))
       .then(function(schema) {
         expect(schema).to.equal(parser.schema);
 
@@ -55,14 +56,15 @@ describe('Schema with $refs to unknown file types', function() {
         // The "circular" flag should NOT be set
         expect(parser.$refs.circular).to.equal(false);
         done();
-      });
+      })
+      .catch(done);
   });
 
   it('should bundle successfully', function(done) {
     testDone = done;
     var parser = new $RefParser();
-    parser
-      .bundle(path.rel('specs/unknown/unknown.yaml'))
+
+    parser.bundle(path.rel('specs/unknown/unknown.yaml'))
       .then(function(schema) {
         expect(schema).to.equal(parser.schema);
 
@@ -70,6 +72,7 @@ describe('Schema with $refs to unknown file types', function() {
         schema.definitions.binary = helper.convertNodeBuffersToPOJOs(schema.definitions.binary);
         expect(schema).to.deep.equal(helper.dereferenced.unknown);
         done();
-      });
+      })
+      .catch(done);
   });
 });